changes made to build without warnings (#143)

+++ src/statespace/dart/MetaSkeletonStateSpace.cpp
@@ -105,7 +105,7 @@ MetaSkeletonPtr MetaSkeletonStateSpace::getMetaSkeleton() const
 void MetaSkeletonStateSpace::convertPositionsToState(
   const Eigen::VectorXd& _positions, State* _state) const
 {
-  if (_positions.size() != mMetaSkeleton->getNumDofs())
+  if (static_cast<size_t>(_positions.size()) != mMetaSkeleton->getNumDofs()) 
     throw std::invalid_argument("Incorrect number of positions.");
 
   for (size_t isubspace = 0; isubspace < getNumSubspaces(); ++isubspace)
@@ -149,7 +149,7 @@ void MetaSkeletonStateSpace::convertStateToPositions(
     subspace->convertStateToPositions(substate, jointPositions);
 
     // TODO: Find a more efficient way to do this mapping.
-    for (size_t idof = 0; idof < jointPositions.size(); ++idof)
+    for (size_t idof = 0; idof < static_cast<size_t>(jointPositions.size()); ++idof) 
     {
       const auto dof = joint->getDof(idof);
       const auto dofIndex = mMetaSkeleton->getIndexOf(dof, false);
